Extracted the Cities, the Treasure Cases and the Stable Objects from the DB.
