+++ src/dg_script/dg_mobcmd.cpp
@@ -33,6 +33,7 @@
 #include "game_skills/townportal.h"
 #include "utils/id_converter.h"
 #include "structs/global_objects.h"
+#include "game_mechanics/stable_objs.h"
 
 struct mob_command_info {
 	const char *command;
@@ -361,7 +362,7 @@ void do_mload(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/, Trigger
 			return;
 		}
 		if (GetObjMIW(object->get_rnum()) >= 0 && obj_proto.actual_count(object->get_rnum()) > GetObjMIW(object->get_rnum())) {
-			if (!IsTimerUnlimited(obj_proto[object->get_rnum()].get())) {
+			if (!stable_objs::IsTimerUnlimited(obj_proto[object->get_rnum()].get())) {
 				sprintf(buf, "mload: Попытка загрузить предмет больше чем в MIW для #%d.", number);
 				mob_log(ch, trig, buf);
 //				extract_obj(object.get());
