+++ src/game_magic/spells.cpp
@@ -26,10 +26,10 @@
 #include "communication/parcel.h"
 #include "administration/privilege.h"
 #include "color.h"
-#include "game_skills/townportal.h"
 #include "cmd/flee.h"
 #include "stuff.h"
 #include "utils/utils_char_obj.inl"
+#include "game_mechanics/stable_objs.h"
 
 #include <cmath>
 
@@ -1561,7 +1561,7 @@ void mort_show_obj_values(const ObjData *obj, CharData *ch, int fullness, bool e
 	SendMsgToChar(CCNRM(ch, C_NRM), ch);
 //enhansed_scroll = true; //для теста
 	if (enhansed_scroll) {
-		if (IsTimerUnlimited(obj))
+		if (stable_objs::IsTimerUnlimited(obj))
 			sprintf(buf2, "Таймер: %d/нерушимо.", obj_proto[GET_OBJ_RNUM(obj)]->get_timer());
 		else
 			sprintf(buf2, "Таймер: %d/%d.", obj_proto[GET_OBJ_RNUM(obj)]->get_timer(), obj->get_timer());
