+++ src/game_quests/bodrich_quests.h
@@ -1,0 +1,14 @@
+//
+// Created by Sventovit on 07.09.2024.
+//
+
+#ifndef BYLINS_SRC_GAME_QUESTS_BODRICH_QUESTS_H_
+#define BYLINS_SRC_GAME_QUESTS_BODRICH_QUESTS_H_
+
+namespace bodrich_quest {
+
+} // namespace bodrich_quest
+
+#endif //BYLINS_SRC_GAME_QUESTS_BODRICH_QUESTS_H_
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
