+++ src/mapsystem.cpp
@@ -10,8 +10,7 @@
 #include "noob.h"
 #include "utils/utils_char_obj.inl"
 #include "game_mechanics/guilds.h"
-
-extern std::vector<City> Cities;
+#include "game_mechanics/cities.h"
 
 namespace Noob {
 int outfit(CharData *ch, void *me, int cmd, char *argument);
@@ -572,15 +571,10 @@ void print_map(CharData *ch, CharData *imm) {
 	MAX_LINES = MAX_LINES_STANDART;
 	MAX_LENGTH = MAX_LENGTH_STANDART;
 	MAX_DEPTH_ROOMS = MAX_DEPTH_ROOM_STANDART;
-	if (ch->map_check_option(MAP_MODE_BIG)) {
-		for (unsigned int i = 0; i < Cities.size(); i++) {
-			if (GetZoneVnumByCharPlace(ch) == Cities[i].rent_vnum / 100) {
-				MAX_LINES = MAX_LINES_BIG;
-				MAX_LENGTH = MAX_LENGTH_BIG;
-				MAX_DEPTH_ROOMS = MAX_DEPTH_ROOM_BIG;
-				break;
-			}
-		}
+	if (ch->map_check_option(MAP_MODE_BIG) && cities::IsCharInCity(ch)) {
+		MAX_LINES = MAX_LINES_BIG;
+		MAX_LENGTH = MAX_LENGTH_BIG;
+		MAX_DEPTH_ROOMS = MAX_DEPTH_ROOM_BIG;
 	}
 	for (unsigned i = 0; i < MAX_LINES; ++i) {
 		for (unsigned k = 0; k < MAX_LENGTH; ++k) {
